refactor: extract AppState issue-export queries extension (#589, #430)
Ninth slice of #430 (AppState decomposition). Per #587 design sketch:
extracts 6 sync passthrough delegators to `issueExportController` — the
issue-export query surface (canExportIssues, canRequestIssueExport,
issueExportSetupMessage, issueExportRoutingMessage,
defaultGitHubIssueExportTarget, defaultJiraIssueExportTarget) — into a
new Sources/BugNarrator/AppState+IssueExportQueries.swift as
`extension AppState { ... }`.

Range: AppState.swift lines 708-730 (23 lines, 1039 bytes, SHA-256
8b5c05f74b4f1af0202925a8da08a28196d8394218096e2d5b915426d7d32e18).
Byte-identical bodies; single-collaborator (issueExportController) domain.

Confirmed via delegator-vs-orchestrator test — all 6 pass:
single-expression bodies, no `await`, no `@Published` mutation. Two
methods (canExportIssues, canRequestIssueExport) pass `status.phase`
as a read-only context argument — proven-delegator pattern per #580's
setExportReviewResolution precedent.

Full triple-proof stack applied:
- Endpoint verification (708 = `func canExportIssues(...) -> Bool {`,
  730 = `}` of defaultJiraIssueExportTarget, 731 = blank)
- SHA gate matched at both spec-review (against origin) and post-extract
  (against new-file body): both hash to 8b5c05f7...
- Byte-diff against origin/main: silent (exit 0)
- Post-deletion double-blank scan: no double-blanks created (line 707
  was already blank; deleting 708-731 including the trailing blank left
  a single-blank seam)

Extension inherits @MainActor from AppState per SE-0316 — NOT
re-annotated. New file imports `import Foundation` (matches house style;
strictly unnecessary per codex, but consistent with all other AppState+*
extensions).

No dead-import cleanup — codex confirmed AppState.swift still needs
Combine + Foundation.

667 unit tests pass unchanged; docs-sync validation passes.

Pre-build spec reviewed by codex in a fresh session (3 findings folded:
Foundation as house-style vs. strictly-required, double-blank collapse
protocol, status.phase-passthrough delegator classification per #580).
