+++ Sources/BugNarrator/AppState.swift
@@ -705,30 +705,6 @@ final class AppState: ObservableObject {
         )
     }
 
-    func canExportIssues(from session: TranscriptSession, to destination: ExportDestination) -> Bool {
-        issueExportController.canExportIssues(from: session, to: destination, statusPhase: status.phase)
-    }
-
-    func canRequestIssueExport(from session: TranscriptSession) -> Bool {
-        issueExportController.canRequestIssueExport(from: session, statusPhase: status.phase)
-    }
-
-    func issueExportSetupMessage(for destination: ExportDestination) -> String? {
-        issueExportController.issueExportSetupMessage(for: destination)
-    }
-
-    func issueExportRoutingMessage(for destination: ExportDestination, session: TranscriptSession) -> String? {
-        issueExportController.issueExportRoutingMessage(for: destination, session: session)
-    }
-
-    func defaultGitHubIssueExportTarget() -> GitHubIssueExportTarget? {
-        issueExportController.defaultGitHubIssueExportTarget()
-    }
-
-    func defaultJiraIssueExportTarget() -> JiraIssueExportTarget? {
-        issueExportController.defaultJiraIssueExportTarget()
-    }
-
     func startSession() async {
         recordingLogger.info(.sessionStartRequested, "A feedback session start was requested.")
